+++ extensions/realtime/extend.php
@@ -70,7 +70,8 @@ return [
     (new Extend\Event)
         ->subscribe(Push\EventSubscriber::class)
         ->listen(\Flarum\Notification\Event\Sent::class, Push\Listener\BroadcastNotifications::class)
-        ->listen(\Flarum\Settings\Event\Saved::class, Listener\RestartServerOnSettingChange::class),
+        ->listen(\Flarum\Settings\Event\Saved::class, Listener\RestartServerOnSettingChange::class)
+        ->listen(\Flarum\Frontend\Event\AssetsRecompiled::class, Listener\BroadcastAssetsRevision::class),
 
     (new Extend\Notification)
         ->driver('realtime', Push\NotificationDriver::class),