+++ extensions/tags/js/dist-typings/admin/components/TagsPage.d.ts
@@ -2,6 +2,7 @@ export default class TagsPage extends ExtensionPage<import("flarum/admin/compone
     constructor();
     oninit(vnode: any): void;
     forcedRefreshKey: number | undefined;
+    sortable: typeof import("sortablejs") | undefined;
     content(): JSX.Element;
     onListOnCreate(vnode: any): void;
     setMinTags(minTags: any, maxTags: any, value: any): void;